Fixed cast from float to unsigned int in Weapon class and removed commented out code lines.

+++ src/Weapon.cpp
@@ -48,9 +48,7 @@ void Weapon::clampToMaxRange(const sf::Vector2u& origin, sf::Vector2u& target) c
 	auto f_target = static_cast<sf::Vector2f>(target);
 	auto path = f_target - f_origin;
 	if (Util::vecLength(path) > (float) range) {
-		//std::cout << "CLAMPED FROM: (" << target.x << ", " << target.y << ")" << std::endl;
 		path = Util::vecUnit(path) * (float) range;
-		target = origin + sf::Vector2u(std::round(path.x), std::round(path.y));
-		//std::cout << "TO: (" << target.x << ", " << target.y << ")" << std::endl;
+		target = origin + sf::Vector2u(static_cast<unsigned int>(std::round(path.x)), static_cast<unsigned int>(std::round(path.y)));
 	}
 }
